linux/x11: Store refresh rate on X11Window itself
